Reject Screensaver with empty overview
I'm not sure whether this is a bug. But I would prefer to reject the found games with empty overview. 

For example, if I'm running:

-lang=es -missing=missing.txt -refresh -use_ss -use_gdb=false -image_dir=downloaded_images -image_path=downloaded_images -workers=1

I would expect it would only refresh the games with a description of the game. In any other case, leave it as it is.

What do you thing?