Add a transcode option.
Because all the videos from screenscraper.fr are too heavy for the poor
raspberry pi, I've added an option to the scraper to transcode them on
download. This will save me lots of time when scraping as I won't have
to manually transcode the videos. Note that this requires HandBrakeCLI
to be installed and it's probably not the best idea to do it on the Pi
itself as it's pretty CPU intensive.